+++ ext_emconf.php
@@ -27,7 +27,7 @@ $EM_CONF[$_EXTKEY] = [
     'author_company' => '',
 
     'state' => 'stable',
-    'version' => '1.1.7',
+    'version' => '1.2.0',
 
     'uploadFolders' => false,
     'createDirs' => '',